refactor: update package structure
