+++ api/pybot/agent/__init__.py
@@ -4,9 +4,17 @@ from langchain.agents import AgentOutputParser
 from langchain_core.agents import AgentAction, AgentActionMessageLog
 from langchain_core.language_models import BaseLanguageModel
 from langchain_core.messages import AIMessage, BaseMessage, SystemMessage
-from langchain_core.prompts import ChatPromptTemplate
+from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
 from langchain_core.runnables import Runnable, RunnablePassthrough
 from langchain_core.tools import BaseTool
+from langchain_openai import ChatOpenAI
+
+from pybot.agent.executor import PybotAgentExecutor
+from pybot.agent.output_parser import MarkdownOutputParser
+from pybot.agent.prompt import SYSTEM
+from pybot.config import settings
+from pybot.memory import memory
+from pybot.tools import CodeSandbox
 
 
 def construct_scratchpad(
@@ -40,3 +48,31 @@ def create_agent(
         | output_parser
     )
     return agent
+
+
+llm = ChatOpenAI(
+    openai_api_base=str(settings.llm.url),
+    model=settings.llm.model,
+    openai_api_key=settings.llm.creds,
+    max_tokens=1024,
+    streaming=True,
+)
+
+prompt = ChatPromptTemplate.from_messages(
+    [
+        ("system", SYSTEM),
+        MessagesPlaceholder(variable_name="history"),
+        ("user", "{input}"),
+        MessagesPlaceholder(variable_name="agent_scratchpad"),
+    ]
+)
+tools = [
+    CodeSandbox(
+        gateway_url=str(settings.jupyter.gateway_url),
+    )
+]
+output_parser = MarkdownOutputParser(language_actions={"python": "python"})
+agent = create_agent(llm, tools, prompt, output_parser)
+agent_executor = PybotAgentExecutor(
+    agent=agent, tools=tools, memory=memory, handle_parsing_errors=True
+)
